requirements.txt
flask==3.0.3
flask-cors==4.0.1
pyyaml==6.0.2
boto3==1.34.69
pyarrow==16.1.0
fastavro==1.9.4
pyiceberg[s3fs,glue]==0.7.1
requests==2.32.3
python-dotenv==1.0.1
croniter==2.0.5
aiosqlite==0.20.0
pymysql==1.1.1
psycopg2-binary==2.9.9
pymongo==4.7.3
pymssql==2.3.0
mysql-replication==1.0.2
hvac==2.3.0
anthropic>=0.30.0
simple-salesforce>=1.12.0
deltalake>=0.17.0
azure-cosmos>=4.6.0
pinotdb>=0.4.0
splunk-sdk>=1.7.4
google-cloud-spanner>=3.40.0
cassandra-driver>=3.29.0
clickhouse-connect>=0.7.0
google-ads>=24.0.0
google-auth>=2.29.0
